pyproject.toml
[project]
name = "molop"
dynamic = ["version"]
description = "Molcule OPerator"
authors = [{ name = "tmj", email = "mj_t@zju.edu.cn" }]
requires-python = ">=3.10" # python 3.10 or later, follow the main dependence of rdkit
readme = "README.md"
license = "MIT"
maintainers = [{ name = "tmj", email = "mj_t@zju.edu.cn" }]
keywords = [
    "cheminformatics",
    "xyz2mol",
    "parser",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fire>=0.7.0",
    "joblib>=1.5.1",
    "openbabel-wheel>=3.1.1.22",
    "pandas>=2.3.1",
    "pint>=0.24.2",
    "pydantic>=2.11.7",
    "rdkit>=2023.9.6",
    "rdkit-dof>=0.1.3",
    "regex<=2025.7.33",
    "scipy>=1.15.3",
    "tabulate>=0.9.0",
    "timeout-decorator>=0.5.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
desc = [
    "mordred~=1.2",
    "dscribe~=2.1",
    "ase~=3.22",
]
full = [
    "mordred~=1.2",
    "dscribe~=2.1",
    "ase~=3.22",
]

[project.urls]
Repository = "https://github.com/gentle1999/MolOP"

[project.scripts]
molop = "molop.cli.cli:app"


[dependency-groups]
dev = [
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.7",
    "mkdocs>=1.6.1",
    "mkdocs-jupyter>=0.25.1",
    "mkdocs-material>=9.7.0",
    "mkdocs-material-extensions>=1.3.1",
    "mkdocs-static-i18n>=1.3.0",
    "mkdocstrings-python>=1.19.0",
    "mypy>=1.18.2",
    "nbconvert>=7.16.6",
    "pydantic[mypy]>=2.12.5",
    "pyright>=1.1.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "rich>=14.2.0",
    "ruff>=0.14.6",
    "seaborn>=0.13.2",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.ruff]
exclude = [".git", ".venv", "__pycache__", "build", "dist", "*.ipynb"]
line-length = 100
indent-width = 4
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",
    "W",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort 
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions 
    "UP",  # pyupgrade 
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long 
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.isort]
lines-after-imports = 2


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["src"]


[tool.mypy]
plugins = ['pydantic.mypy']
warn_return_any = false        # 允许函数返回 Any (这对调用未加类型的第三方库很有用)
warn_unused_configs = true     # 警告无用的配置
explicit_package_bases = true  # 配合 src 目录结构

disallow_untyped_defs = false  # 关键！允许编写没有类型注解的函数
disallow_any_generics = false  # 允许使用 List, Dict 而不必须写 List[int]

check_untyped_defs = true      # 关键！即使函数没写注解，也要检查函数内部的逻辑错误
no_implicit_optional = true    # 禁止隐式的 Optional (即禁止 x: str = None)
warn_redundant_casts = true    # 警告多余的 cast
warn_unused_ignores = false     # 警告多余的 # type: ignore

ignore_missing_imports = true
exclude = ["tests", "scripts", "build", "dist"]


[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true


[[tool.uv.index]]
name = "pypi"
url = "https://mirrors.zju.edu.cn/pypi/web/simple"

[[tool.uv.index]]
name = "hong"
url = "http://10.72.201.58:13000/api/packages/tmj/pypi/simple/"
publish-url = "http://10.72.201.58:13000/api/packages/tmj/pypi/"
explicit = true
